% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpful.functions.ipdma.R
\name{appendix_example_continuous}
\alias{appendix_example_continuous}
\title{Generate a fake IPD-MA data for demonstration}
\usage{
appendix_example_continuous()
}
\description{
Generate a fake IPD-MA data for demonstration
}
