% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.related.functions.R
\name{findMissingPattern}
\alias{findMissingPattern}
\title{Find missing data pattern, i.e. whether they are systematically missing or sporadically missing}
\usage{
findMissingPattern(dummydata, covariates, studyname = "study")
}
\arguments{
\item{dummydata}{Dataset which contains covariates of interest}

\item{covariates}{A vector of covariate names to find missing data pattern}
}
\description{
Find missing data pattern in a given dataset
}
