% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.example.data.R
\name{generate_ipdma_example}
\alias{generate_ipdma_example}
\title{Generate a fake IPD-MA data for demonstration}
\usage{
generate_ipdma_example(type = "continuous")
}
\arguments{
\item{type}{"continuous" for continuous outcome and "binary" for binary outcome}
}
\description{
Generate a fake IPD-MA data for demonstration
}
