% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpful.functions.R
\name{generate_ipdma_example_with_systematic_missing}
\alias{generate_ipdma_example_with_systematic_missing}
\title{Generate a fake IPD-MA data with systematically missing variables for demonstration}
\usage{
generate_ipdma_example_with_systematic_missing()
}
\description{
Generate a fake IPD-MA data with systematically missing variables for demonstration
}
