% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.example.data.R
\name{generate_ipdnma_example}
\alias{generate_ipdnma_example}
\title{Generate a simualted IPD-NMA data for demonstration}
\usage{
generate_ipdnma_example(type = "continuous")
}
\arguments{
\item{type}{"continuous" for continuous outcome and "binary" for binary outcome}
}
\value{
return simulated IPD-NMA data
ds <- generate_ipdnma_example(type = "continuous")
head(ds)
}
\description{
Generate a simulated IPD-NMA data for demonstration
}
