% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.related.functions.R
\name{getCorrectMeth}
\alias{getCorrectMeth}
\title{Find correct imputation method to be used in the mice package}
\usage{
getCorrectMeth(
  dataset,
  missingPattern,
  typeofvar = NULL,
  studyname = "study",
  treatmentname = "treat",
  interaction = FALSE
)
}
\arguments{
\item{dataset}{data which contains variables of interest}

\item{missingPattern}{missing pattern object created using \code{\link{findMissingPattern}}.}

\item{typeofvar}{type of variable: continuous, binary, or count}

\item{studyname}{study name}

\item{treatmentname}{treatment name}

\item{interaction}{whether to include interaction or not}
}
\description{
Find correct imputation method for the mice package
}
