% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.related.functions.R
\name{getCorrectPred}
\alias{getCorrectPred}
\title{Find correct imputation prediction matrix to be used in the mice package}
\usage{
getCorrectPred(
  dataset,
  missingPattern,
  studyname = "study",
  treatmentname = "treat",
  outcomename = "y",
  interaction = FALSE
)
}
\arguments{
\item{dataset}{data which contains variables of interest}

\item{missingPattern}{missing pattern object created using \code{\link{findMissingPattern}}}

\item{studyname}{study name}

\item{treatmentname}{treatment name}

\item{outcomename}{outcome name}

\item{interaction}{whether to include interaction or not}
}
\description{
Find correct imputation prediction matrix for the mice package
}
