% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipd.summary.R
\name{ipd.gelman.plot}
\alias{ipd.gelman.plot}
\title{Use coda package to plot Gelman-Rubin diagnostic plot}
\usage{
ipd.gelman.plot(result)
}
\arguments{
\item{result}{Object created by \code{\link{ipd.run}} function}
}
\value{
None
}
\description{
This function plots Gelman-Rubin diagnostic using coda package.
}
