% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipd.model.R
\name{ipd.model}
\alias{ipd.model}
\title{Make an ipd object containing data, priors, and a JAGS model file}
\usage{
ipd.model(
  y = NULL,
  study = NULL,
  treat = NULL,
  X = NULL,
  response = "continuous",
  model = "onestage",
  shrinkage = "none"
)
}
\arguments{
\item{y}{outcome of the study. Can be continuous or binary.}

\item{study}{a vector indicating which study the patient belongs to. Please change the study names into numbers (i.e. 1,2,3,etc)}

\item{treat}{a vector indicating which treatment the patient was assigned to. Please change the study names into numbers (i.e. 1,2,3,etc)}

\item{X}{a matrix of covariate values for each patient. Dimension would be number of patients x number of covariates.}

\item{response}{"continuous" for continuous outcome and "binary" for binary outcome}

\item{model}{"onestage" for one-stage model; two-stage are planned to be implemented}

\item{shrinkage}{shrinkage method applied to the effect modifiers. "none" correspond to no shrinkage.
"laplace" corresponds to a adaptive shrinkage with a Laplacian prior (ie often known as Bayesian LASSO).
"SSVS" corresponds to the Search Variable Selection method. SSVS is not strictly a shrinkage method, 
but shrinks the estimated coefficient through variable selection in each iteration of the MCMC. 
See O'hara et al (2009) for more reference.}
}
\value{
\item{data}{Data organized in a list so that it can be used when running code in JAGS}
\item{code}{JAGS code that is used to run the model. Use cat(code) to see the code in a nicer format.}
}
\description{
This function sets up data and JAGS code that is needed to run ipd models in JAGS
}
