% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipd.run.R
\name{ipd.run.parallel}
\alias{ipd.run.parallel}
\title{Run the model using the ipd object with parallel computation}
\usage{
ipd.run.parallel(
  ipd,
  pars.save = c("beta", "gamma", "delta"),
  inits = NULL,
  n.chains = 3,
  n.adapt = 1000,
  n.iter = 10000
)
}
\arguments{
\item{ipd}{ipd object created from \code{\link{ipd.data}} function}

\item{pars.save}{parameters to save. "beta" - coefficients for main effects; "gamma" - coefficients for effect modifiers; "delta" - average treaetment effect}

\item{inits}{initial values specified for the parameters to save}

\item{n.chains}{Number of MCMC chains to}

\item{n.adapt}{The number of iterations for adaptation (this corresponds to the number of burnin)}

\item{n.iter}{The number of iterations to run after the adaptation}

\item{n.cores}{The number of cores used for parallel computation}
}
\description{
This function runs the model through parallel computation from dclone package. Before running this function, we need to specify data, prior, JAGS code, etc. using \code{\link{ipd.data}}.
}
