% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipd.run.R
\name{ipd.run}
\alias{ipd.run}
\title{Run the model using the ipd object}
\usage{
ipd.run(ipd)
}
\arguments{
\item{ipd}{ipd object created from \code{\link{ipd.data}} function}
}
\description{
This is the core function that runs the model in our program. Before running this function, we need to specify data, prior, JAGS code, etc. using \code{\link{ipd.data}}.
}
