% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipdma.impute.R
\name{ipdma.impute}
\alias{ipdma.impute}
\title{Impute missing data in individual participant data with study-level cluster.}
\usage{
ipdma.impute(
  dataset = NULL,
  covariates = NULL,
  typeofvar = NULL,
  interaction = TRUE,
  meth = NULL,
  pred = NULL,
  studyname = NULL,
  treatmentname = NULL,
  outcomename = NULL
)
}
\arguments{
\item{dataset}{Data which contains variables of interests}

\item{covariates}{Vector of variable names to find missing data pattern}

\item{typeofvar}{Type of variables; should be a vector of these values: "continuous", "binary", or "count". Index value of the vector should be predictor names.}

\item{interaction}{Indicator denoting whether treatment-covariate interactions should be included}

\item{meth}{User can specify imputation method to be used in the mice package. If left unspecified, function picks a reasonable one.}

\item{pred}{User can specify correct prediction matrix to be used in the mice package. If left unspecified, function picks a reasonable one.}

\item{studyname}{Study name in the data specified. Default is "study".}

\item{treatmentname}{Treatment name in the data specified. Default is "treat".}

\item{outcomename}{Outcome name in the data specified. Default is "y".}
}
\value{
\item{missingPattern}{missing Pattern object returned by running \code{\link{findMissingPattern}}}
\item{meth}{imputation method used with the mice function}
\item{pred}{prediction matrix used with the mice function}
}
\description{
Impute missing data in individual participant data where data is clustered by different studies.
In the presence of systematically missing variables, we use 2l.2stage.norm, 2l.2stage.bin, 
and 2l.2stage.pois methods depending on the type of the covariates to impute. 
If there are no systematically missing variables, the function defaults to use 2l.pmm in miceadds package 
which generalizes predictive mean matching using linear mixed model.
}
