% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipdma.onestage.model.R
\name{ipdma.model.onestage}
\alias{ipdma.model.onestage}
\title{Make an onestage individual patient data meta analysis (ipd-ma) object containing data, priors, and a JAGS model code}
\usage{
ipdma.model.onestage(
  y = NULL,
  study = NULL,
  treat = NULL,
  X = NULL,
  response = "normal",
  type = "random",
  approach = "deft",
  shrinkage = "none",
  scale = TRUE,
  mean.a = 0,
  prec.a = 0.001,
  mean.beta = 0,
  prec.beta = 0.001,
  mean.gamma = 0,
  prec.gamma = 0.001,
  mean.gamA = 0,
  prec.gamA = 0.001,
  mean.delta = 0,
  prec.delta = 0.001,
  hy.prior = list("dhnorm", 0, 1),
  lambda.prior = NULL,
  p.ind = NULL,
  g = NULL,
  hy.prior.eta = NULL
)
}
\arguments{
\item{y}{outcome of the study. Can be continuous or binary.}

\item{study}{a vector indicating which study the patient belongs to. Please change the study names into numbers (i.e. 1,2,3,etc)}

\item{treat}{a vector indicating which treatment the patient was assigned to (ie 1 for treatment, 0 for placebo)}

\item{X}{a matrix of covariate values for each patient. Dimension would be number of patients x number of covariates.}

\item{response}{Specification of the outcomes type. Must specify either "normal" or "binomial"}

\item{type}{Assumption on the treatment effect: either "random" for random effects model or "fixed" for fixed effects model. Default is "random".}

\item{approach}{"deluded" approach which does not separate within-study information and across-study information or "deft" approach which separate within-study and across-study information by centering the covariates using study specific mean and including study specific mean in the regression. Refer to Fisher et al. for more details.}

\item{shrinkage}{shrinkage method applied to the effect modifiers. "none" correspond to no shrinkage.
"laplace" corresponds to a adaptive shrinkage with a Laplacian prior (ie often known as Bayesian LASSO).
"SSVS" corresponds to the Search Variable Selection method. SSVS is not strictly a shrinkage method, 
but pulls the estimated coefficient toward zero through variable selection in each iteration of the MCMC. 
See O'hara et al (2009) for more details.}

\item{scale}{indicator for scaling the covariates; default is TRUE.}

\item{mean.a}{Prior mean for the study intercept.}

\item{prec.a}{Prior precision for the study intercept}

\item{mean.beta}{Prior mean for the regression coefficients of the main effects of the covariates}

\item{prec.beta}{Prior precision for the regression coefficients of the main effects of the covariates}

\item{mean.gamma}{Prior mean for the effect modifiers. This parameter is not used if penalization is placed on effect modifiers.}

\item{prec.gamma}{Prior precision for the effect modifiers.}

\item{mean.gamA}{Prior mean for the effect modifiers of across study information}

\item{prec.gamA}{prior precision for the effect modifiers of across study information}

\item{mean.delta}{Prior mean for the average treatment effect}

\item{prec.delta}{Prior precision for the average treatment effect}

\item{hy.prior}{Prior for the heterogeneity parameter. Supports uniform, gamma, and half normal for normal and binomial response
It should be a list of length 3, where first element should be the distribution (one of dunif, dgamma, dhnorm) and the next two are the parameters associated with the distribution. For example, list("dunif", 0, 5) give uniform prior with lower bound 0 and upper bound 5 for the heterogeneity parameter.}

\item{lambda.prior}{(Only for shrinkage = "laplace") Two options for laplace shrinkage. We can put a gamma prior on the lambda (i.e. list("dgamma",2,0.1)) or put a uniform prior on the inverse of lambda (i.e. list("dunif",0,5))}

\item{p.ind}{(Only for shrinkage = "SSVS") Prior probability of including each of the effect modifiers. Length should be same as the total length of the covariates.}

\item{g}{(Only for shrinkage = "SSVS") Multiplier for the precision of spike. Default is g = 1000.}

\item{hy.prior.eta}{(Only for shrinkage = "SSVS") Standard deviation of the slab prior. Currently only support uniform distribution. Default is list("dunif", 0, 5)}
}
\value{
\item{data.JAGS}{Data organized in a list so that it can be used when running code in JAGS}
\item{code}{JAGS code that is used to run the model. Use cat(code) to see the code in a readable format}
\item{model.JAGS}{JAGS code in a function. This is used when running model in parallel}
\item{scale.mean}{mean used in scaling covariates}
\item{scale.sd}{standard deviation used in scaling covariates}
\item{Xbar}{study specific average of covariates; used in "deft" approach}
}
\description{
This function sets up data and JAGS code that is needed to run onestage ipd-ma models in JAGS.
}
