% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipdnma.twostage.model.R
\name{ipdnma.model.twostage.first}
\alias{ipdnma.model.twostage.first}
\title{Make a first part of the two-stage individual patient data network meta-analysis object containing data, priors, and a JAGS model code}
\usage{
ipdnma.model.twostage.first(
  y = NULL,
  treat = NULL,
  X = NULL,
  response = "normal",
  shrinkage = "none",
  scale = TRUE,
  mean.a = 0,
  prec.a = 0.001,
  mean.b = 0,
  prec.b = 0.001,
  mean.c = 0,
  prec.c = 0.001,
  mean.d = 0,
  prec.d = 0.001,
  lambda.prior = NULL,
  p.ind = NULL,
  g = NULL,
  hy.prior.eta = NULL
)
}
\arguments{
\item{y}{outcome of the study. Can be continuous or binary.}

\item{treat}{vector indicating which treatment the patient was assigned to. Since this is a network meta-analysis and there would be more than 2 treatments.
However this doesn't mean all studies have more than 2 treatments and they can only have two. One thing to be careful is treatment names should be same for all studies.
Treatment that is assigned 1 would be the baseline treatment.}

\item{X}{matrix of covariate values for each patient. Dimension would be number of patients x number of covariates.}

\item{response}{specification of the outcome type. Must specify either "normal" or "binomial".}

\item{shrinkage}{shrinkage method applied to the effect modifiers. "none" correspond to no shrinkage.
"laplace" corresponds to a adaptive shrinkage with a Laplacian prior (ie often known as Bayesian LASSO).
"SSVS" corresponds to the Stochastic Search Variable Selection method. SSVS is not strictly a shrinkage method, but pulls the estimated coefficient toward zero through variable selection in each iteration of the MCMC. 
See O'hara et al (2009) for more details.}

\item{scale}{indicator for scaling the covariates by the overall average; default is TRUE.}

\item{mean.a}{prior mean for the study intercept}

\item{prec.a}{prior precision for the study intercept}

\item{mean.b}{prior mean for the regression coefficients of the main effects of the covariates; main effects are assumed to have common effect.}

\item{prec.b}{prior precision for the regression coefficients of the main effects of the covariates}

\item{mean.c}{prior mean for the effect modifiers. This parameter is not used if penalization is placed on effect modifiers.}

\item{prec.c}{prior precision for the effect modifiers. This parameter is not used if penalization is placed on effect modifiers.}

\item{mean.d}{prior mean for the average treatment effect}

\item{prec.d}{prior precision for the average treatment effect}

\item{lambda.prior}{(only for shrinkage = "laplace") two options for laplace shrinkage. We can put a gamma prior on the lambda (i.e. list("dgamma",2,0.1)) or put a uniform prior on the inverse of lambda (i.e. list("dunif",0,5))}

\item{p.ind}{(only for shrinkage = "SSVS") prior probability of including each of the effect modifiers. Length should be same as the total length of the covariates.}

\item{g}{(only for shrinkage = "SSVS") multiplier for the precision of spike. Default is g = 1000.}

\item{hy.prior.eta}{(only for shrinkage = "SSVS") standard deviation of the slab prior. Currently only support uniform distribution. Default is list("dunif", 0, 5)}
}
\value{
\item{data.JAGS}{data organized in a list so that it can be used when running code in JAGS}
\item{code}{JAGS code that is used to run the model. Use cat(code) to see the code in a readable format}
\item{model.JAGS}{JAGS code in a function. This is used when running model in parallel}
\item{scale.mean}{mean used in scaling covariates}
\item{scale.sd}{standard deviation used in scaling covariates}
\item{treat.original}{original treatment names from user}
\item{treat.relabeled}{treatment relabeled for the purpose of modelling}
}
\description{
This is the first part of the two-stage IPD-NMA model that sets up data and JAGS code needed. 
Each study is fitted using this function.
}
\examples{
ds <- generate_ipdnma_example(type = "continuous")
ds1 <- ds[ds$studyid == 1,]
ipd1 <- with(ds1, ipdnma.model.twostage.first(y = y, treat = treat, X = cbind(z1, z2), response = "normal", shrinkage = "none"))
\donttest{
samples1 <- ipd.run(ipd1)
}
}
\references{
Dias S, Sutton AJ, Ades AE, et al. A Generalized Linear Modeling Framework for Pairwise and Network Meta-analysis of Randomized Controlled Trials. \emph{Medical Decision Making}. 2013;33(5):607-617. \doi{10.1177/0272989X12458724}

O'Hara RB, Sillanpaa MJ. A review of Bayesian variable selection methods: what, how and which. \emph{Bayesian Anal}. 2009;4(1):85-117. \doi{10.1214/09-BA403}

Seo M, White IR, Furukawa TA, et al. Comparing methods for estimating patient-specific treatment effects in individual patient data meta-analysis. \emph{Stat Med}. 2021;40(6):1553-1573. \doi{10.1002/sim.8859}
}
