% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipdnma.twostage.model.R
\name{ipdnma.model.twostage.second}
\alias{ipdnma.model.twostage.second}
\title{Make the second part of the two-stage individual patient data network meta-analysis object containing data, priors, and a JAGS model code}
\usage{
ipdnma.model.twostage.second(
  ipd_all = NULL,
  samples_all = NULL,
  mean.a = 0,
  prec.a = 0.001,
  mean.b = 0,
  prec.b = 0.001,
  mean.c = 0,
  prec.c = 0.001,
  mean.d = 0,
  prec.d = 0.001,
  lambda.prior = NULL,
  p.ind = NULL,
  g = NULL,
  hy.prior.eta = NULL
)
}
\arguments{
\item{ipd_all}{list of first part of the two-stage IPD-NMA model object}

\item{samples_all}{list of samples from running first part of the two-stage IPD-NMA model}

\item{mean.a}{prior mean for the study intercept}

\item{prec.a}{prior precision for the study intercept}

\item{mean.b}{prior mean for the regression coefficients of the main effects of the covariates; main effects are assumed to have common effect.}

\item{prec.b}{prior precision for the regression coefficients of the main effects of the covariates}

\item{mean.c}{prior mean for the effect modifiers. This parameter is not used if penalization is placed on effect modifiers.}

\item{prec.c}{prior precision for the effect modifiers. This parameter is not used if penalization is placed on effect modifiers.}

\item{mean.d}{prior mean for the average treatment effect}

\item{prec.d}{prior precision for the average treatment effect}
}
\value{
\item{data.JAGS}{data organized in a list so that it can be used when running code in JAGS}
\item{code}{JAGS code that is used to run the model. Use cat(code) to see the code in a readable format}
\item{model.JAGS}{JAGS code in a function. This is used when running model in parallel}
}
\description{
This is the second part of the two-stage IPD-NMA model that sets up data and JAGS code needed. 
This model aggregates results from first stage model.
}
\examples{
ds <- generate_ipdnma_example(type = "continuous")
#TODO
}
\references{
Dias S, Sutton AJ, Ades AE, et al. A Generalized Linear Modeling Framework for Pairwise and Network Meta-analysis of Randomized Controlled Trials. \emph{Medical Decision Making}. 2013;33(5):607-617. \doi{10.1177/0272989X12458724}

Seo M, White IR, Furukawa TA, et al. Comparing methods for estimating patient-specific treatment effects in individual patient data meta-analysis. \emph{Stat Med}. 2021;40(6):1553-1573. \doi{10.1002/sim.8859}
}
