% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipdma.impute.R
\name{preprocess.data}
\alias{preprocess.data}
\title{Function to preprocess dataset}
\usage{
preprocess.data(
  dataset = NULL,
  covariates = NULL,
  typeofvar = NULL,
  interaction = NULL,
  studyname = NULL,
  treatmentname = NULL,
  outcomename = NULL
)
}
\arguments{
\item{dataset}{Data which contains variables of interests}

\item{covariates}{Vector of variable names that the user is interested in finding a missing data pattern}

\item{typeofvar}{Type of covariate variables; should be a vector of these values: "continuous", "binary", or "count". Order should follow that of covariates parameter.}

\item{interaction}{Indicator denoting whether treatment-covariate interactions should be included. Default is set to true.}

\item{studyname}{Study name in the data specified.}

\item{treatmentname}{Treatment name in the data specified.}

\item{outcomename}{Outcome name in the data specified.}
}
\description{
Function to preprocess dataset
}
