% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpful.functions.R
\name{treatment.effect}
\alias{treatment.effect}
\title{Calculate patient-specific treatment effect}
\usage{
treatment.effect(
  ipd = NULL,
  samples = NULL,
  newpatient = NULL,
  reference = NULL,
  quantile = c(0.025, 0.5, 0.975)
)
}
\arguments{
\item{ipd}{ipd object created from ipd.data function}

\item{samples}{mcmc samples found from running `ipd.run`}

\item{newpatient}{Covariate values of patients that you want to predict treatment effect on. Must have length equal to total number of covariates.}

\item{reference}{reference group used for finding patient-specific treatment effect; only used for "deft" approach}

\item{quantile}{quantile for the confidence interval}
}
\description{
Convenient function for calculating the patient-specific treatment effect.
Patient-specific treatment effect includes the main effect of treatment and 
treatment-covariate interaction effect (i.e. effect modification). 
Reports odds ratio for the binary outcome.
}
\references{
Riley RD, Debray TP, Fisher D, et al. Individual participant data meta-analysis to examine interactions between treatment effect and participant-level covariates: Statistical recommendations for conduct and planning. \emph{Stat Med}. 2020:39(15):2115-2137. [\url{https://doi.org/10.1002/sim.8516}
}
