% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bipd.R
\docType{package}
\name{bipd-package}
\alias{bipd-package}
\title{bipd: A package for individual patient data (network) meta analysis using Bayesian methods}
\description{
A package for running Bayesian individual patient data (network) meta analysis
}
\references{
A.J. Franchini, S. Dias, A.E. Ades, J.P. Jansen, N.J. Welton (2012), \emph{Accounting for correlation in network meta-analysis with multi-arm trials}, Research Synthesis Methods 3(2):142-160. [\url{https://doi.org/10.1002/jrsm.1049}]
}
\seealso{
\code{\link{ipd.data}}, \code{\link{ipd.run}}
}
