% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bipd.R
\docType{package}
\name{bipd-package}
\alias{bipd-package}
\title{bipd: A package for individual patient data (network) meta analysis using Bayesian methods}
\description{
A package for running Bayesian individual patient data (network) meta analysis
}
