% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bipd.R
\docType{package}
\name{bipd-package}
\alias{bipd-package}
\title{bipd: A package for individual patient data meta-analysis using 'JAGS'}
\description{
A package for individual patient data meta-analysis using 'JAGS'
}
\details{
We use a Bayesian approach to run individual patient data meta-analysis using 'JAGS'. 
The methods incorporate shrinkage methods and calculate patient-specific treatment effects as described in Seo et al. (2021) <DOI:10.1002/sim.8859>. 
This package also includes user-friendly functions that impute missing data in an individual patient data using mice-related packages.
}
\references{
Audigier V, White I, Jolani S, et al. Multiple Imputation for Multilevel Data with Continuous and Binary Variables. \emph{Statistical Science}. 2018;33(2):160-183. \doi{10.1214/18-STS646}

Debray TPA, Moons KGM, Valkenhoef G, et al. Get real in individual participant data (IPD) meta-analysis: a review of the methodology. \emph{Res Synth Methods}. 2015;6(4):293-309. \doi{10.1002/jrsm.1160}

Fisher DJ, Carpenter JR, Morris TP, et al. Meta-analytical methods to identify who benefits most from treatments: daft, deluded, or deft approach?. \emph{BMJ}. 2017;356:j573 \doi{10.1136/bmj.j573}

Riley RD, Debray TP, Fisher D, et al. Individual participant data meta-analysis to examine interactions between treatment effect and participant-level covariates: Statistical recommendations for conduct and planning. \emph{Stat Med}. 2020:39(15):2115-2137. \doi{10.1002/sim.8516}

Seo M, White IR, Furukawa TA, et al. Comparing methods for estimating patient-specific treatment effects in individual patient data meta-analysis. \emph{Stat Med}. 2021;40(6):1553-1573. \doi{10.1002/sim.8859}
}
