% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.related.functions.R
\name{findMissingPattern}
\alias{findMissingPattern}
\title{Find missing data pattern in a given data}
\usage{
findMissingPattern(dataset, covariates, studyname = "study")
}
\arguments{
\item{dataset}{data which contains variables of interest}

\item{covariates}{vector of variable names to find missing data pattern}

\item{studyname}{study name}
}
\description{
Find missing data pattern in a given data i.e. whether they are systematically missing or sporadically missing
}
