% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpful.functions.R
\name{findPerformance}
\alias{findPerformance}
\title{Find performance metrics}
\usage{
findPerformance(
  testoutcome = NULL,
  predictions = NULL,
  aggregation = "weighted"
)
}
\arguments{
\item{testoutcome}{Outcome of the testing data to calculate performance metrics}

\item{predictions}{Predictions from model developed}

\item{aggregation}{Aggregation method to be used to combine performance metrics across studies. 
There are three options: "simple" gives you a simple average across studies; 
"weighted" gives you a weighted average taking into account the sample size of each study. 
"ignore" gives you performance metrics ignoring study level. Default is set to "weighted".}
}
\value{
\item{performancemetrics}{A vector of MSE, MAE, and R-squared values}
}
\description{
Find performance metrics (MSE, MAE, R-squared) taking into account the study level.
}
