% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.example.data.R
\name{generate_sysmiss_ipdma_example}
\alias{generate_sysmiss_ipdma_example}
\title{Generate a fake IPD-MA data with systematically missing covariates}
\usage{
generate_sysmiss_ipdma_example(
  Nstudies = 10,
  Ncov = 5,
  sys_missing_prob = 0.3,
  signal = "small",
  interaction = FALSE,
  aggregation_bias = FALSE
)
}
\arguments{
\item{Nstudies}{Number of studies. Default is 10.}

\item{Ncov}{Number of covariates in total. Options are 5 or 10 studies. Default is set to 5.}

\item{sys_missing_prob}{Proportion of systematically missing studies for each covariate. Default is set to 0.3.}

\item{signal}{Signal to noise ratio between predictors and the outcome by changing the variance of the random error. Options are "small" or "large".}

\item{interaction}{Whether to include treatment indicator and treatment}

\item{aggregation_bias}{Induce aggregation bias in each predictors}
}
\description{
Generate a fake IPD-MA data with systematically missing covariates
}
