% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.related.functions.R
\name{getCorrectMeth}
\alias{getCorrectMeth}
\title{Find correct imputation method to be used in the mice package}
\usage{
getCorrectMeth(
  dataset,
  missingPattern,
  studyname = "study",
  treatmentname = "treat",
  outcomename = "y",
  interaction = FALSE,
  typeofvar = NULL
)
}
\arguments{
\item{dataset}{data which contains variables of interest}

\item{missingPattern}{missing pattern object created using \code{\link{findMissingPattern}}}

\item{studyname}{study name}

\item{treatmentname}{treatment name}

\item{outcomename}{outcome name}

\item{interaction}{indicator for including covariate-treatment interactions}

\item{typeofvar}{type of variables; should be a vector of these values: "continuous", "binary", or "count". Index value of the vector should be predictor names.}
}
\description{
Find correct imputation method for the mice package
}
