% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipdma.impute.R
\name{getCorrectMeth}
\alias{getCorrectMeth}
\title{Find correct imputation method to be used in the mice package}
\usage{
getCorrectMeth(
  dataset = NULL,
  missingPattern = NULL,
  typeofvar = NULL,
  interaction = TRUE,
  studyname = NULL,
  treatmentname = NULL,
  outcomename = NULL
)
}
\arguments{
\item{dataset}{data which contains variables of interest}

\item{missingPattern}{missing pattern object created using \code{\link{findMissingPattern}}}

\item{typeofvar}{type of variables; should be a vector of these values: "continuous", "binary", or "count". Index value of the vector should be predictor names.}

\item{interaction}{indicator for including covariate-treatment interactions}

\item{studyname}{study name}

\item{treatmentname}{treatment name}

\item{outcomename}{outcome name}
}
\description{
Find correct imputation method for the mice package
}
