% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipdma.impute.R
\name{getCorrectMeth}
\alias{getCorrectMeth}
\title{Find correct imputation method to be used in the mice package}
\usage{
getCorrectMeth(
  dataset = NULL,
  missingPattern = NULL,
  sys_impute_method = "2l.2stage",
  interaction = TRUE
)
}
\arguments{
\item{dataset}{data which contains variables of interest}

\item{missingPattern}{missing pattern object created using \code{\link{findMissingPattern}}}

\item{sys_impute_method}{Method used for systematically missing studies. Options are "2l.glm", 2l.2stage", or "2l.jomo". Default is set to "2l.2stage".
For more details, Read micemd package for suggestions on which method to use depending on observed clusters and observed values per cluster.
There is also last option where you simply ignore all the clustering level and impute using predictive mean matching. To specify such option, set this parameter to "pmm".}

\item{interaction}{Indicator denoting whether treatment-covariate interactions should be included. Default is set to true.#'}
}
\description{
Find correct imputation method for the mice package
}
