% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipdma.impute.R
\name{getCorrectMeth}
\alias{getCorrectMeth}
\title{Find correct imputation method to be used in the mice package}
\usage{
getCorrectMeth(dataset = NULL, missingPattern = NULL, interaction = TRUE)
}
\arguments{
\item{dataset}{data which contains variables of interest}

\item{missingPattern}{missing pattern object created using \code{\link{findMissingPattern}}}

\item{interaction}{indicator for including covariate-treatment interactions}
}
\description{
Find correct imputation method for the mice package
}
