% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipdma.impute.R
\name{getCorrectPred}
\alias{getCorrectPred}
\title{Find correct imputation prediction matrix to be used in the mice package}
\usage{
getCorrectPred(dataset = NULL, missingPattern = NULL, interaction = TRUE)
}
\arguments{
\item{dataset}{data which contains variables of interest}

\item{missingPattern}{missing pattern object created using \code{\link{findMissingPattern}}}

\item{interaction}{Indicator denoting whether treatment-covariate interactions should be included. Default is set to true.#'}
}
\description{
Find correct imputation prediction matrix for the mice package
}
