% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipd.model.R
\name{ipd.model}
\alias{ipd.model}
\title{Make an ipd object containing data, priors, and a JAGS model file}
\usage{
ipd.model(
  y = NULL,
  study = NULL,
  treat = NULL,
  X = NULL,
  response = "normal",
  type = "random",
  model = "onestage",
  shrinkage = "none",
  mean.a = 0,
  prec.a = 0.001,
  mean.beta = 0,
  prec.beta = 0.001,
  mean.gamma = 0,
  prec.gamma = 0.001,
  mean.delta = 0,
  prec.delta = 0.001,
  hy.prior = list("dhnorm", 0, 1),
  lambda.prior = NULL,
  p.ind = NULL
)
}
\arguments{
\item{y}{outcome of the study. Can be continuous or binary.}

\item{study}{a vector indicating which study the patient belongs to. Please change the study names into numbers (i.e. 1,2,3,etc)}

\item{treat}{a vector indicating which treatment the patient was assigned to (ie 1 for treatment, 0 for placebo)}

\item{X}{a matrix of covariate values for each patient. Dimension would be number of patients x number of covariates.}

\item{response}{Specification of the outcomes type. Must specify either "normal" or "binomial"}

\item{type}{Assumption on the treatment effect: either "random" for random effects model or "fixed" for fixed effects model. Default is "random".}

\item{model}{"onestage" for one-stage model; two-stage are planned to be implemented}

\item{shrinkage}{shrinkage method applied to the effect modifiers. "none" correspond to no shrinkage.
"laplace" corresponds to a adaptive shrinkage with a Laplacian prior (ie often known as Bayesian LASSO).
"SSVS" corresponds to the Search Variable Selection method. SSVS is not strictly a shrinkage method, 
but pulls the estimated coefficient toward zero through variable selection in each iteration of the MCMC. 
See O'hara et al (2009) for more reference.}

\item{mean.a}{Prior mean for the study intercept}

\item{prec.a}{Prior precision for the study intercept}

\item{mean.beta}{Prior mean for the regression coefficients of the main effects of the covariates}

\item{prec.beta}{Prior precision for the regression coefficients of the main effects of the covariates}

\item{mean.gamma}{Prior mean for the effect modifiers. This parameter is not used if penalization is placed on effect modifiers.}

\item{prec.gamma}{Prior precision for the effect modifiers.}

\item{mean.delta}{Prior mean for the treatment effect}

\item{prec.delta}{Prior precision for the treatment effect}

\item{hy.prior}{Prior for the heterogeneity parameter. Supports uniform, gamma, and half normal for normal and binomial response
It should be a list of length 3, where first element should be the distribution (one of dunif, dgamma, dhnorm) and the next two are the parameters associated with the distribution. For example, list("dunif", 0, 5) give uniform prior with lower bound 0 and upper bound 5 for the heterogeneity parameter.}

\item{lambda.prior}{(Only for shrinkage = "laplace") Two options for laplace shrinkage. We can put a gamma prior on the lambda (i.e. list("dgamma",2,0.1)) or put a uniform prior on the inverse of lambda (i.e. list("dunif",0,5))}

\item{p.ind}{(Only for shrinkage = "SSVS") Prior probability of including each of the effect modifiers. Length should be same as the total length of the covariates.}
}
\value{
\item{data.JAGS}{Data organized in a list so that it can be used when running code in JAGS}
\item{code}{JAGS code that is used to run the model. Use cat(code) to see the code in a nicer format.}
\item{code2}{JAGS code in a function. This is used when running model in parallel}
}
\description{
This function sets up data and JAGS code that is needed to run ipd models in JAGS
}
