% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipd.run.R
\name{ipd.run}
\alias{ipd.run}
\title{Run the model using the ipd object}
\usage{
ipd.run(
  ipd,
  pars.save = c("beta", "gamma", "delta"),
  inits = NULL,
  n.chains = 3,
  n.adapt = 1000,
  n.burnin = 1000,
  n.iter = 10000
)
}
\arguments{
\item{ipd}{ipd object created from ipd.data function}

\item{pars.save}{parameters to save. "beta" - coefficients for main effects; "gamma" - coefficients for effect modifiers; "delta" - average treaetment effect}

\item{inits}{initial values specified for the parameters to save}

\item{n.chains}{Number of MCMC chains to sample}

\item{n.adapt}{The number of iterations for adaptation (Note that the samples from adaptation phase is non-Markovian and do not constitute a Markov chain)}

\item{n.burnin}{The number of iterations for burn-in}

\item{n.iter}{The number of iterations to run after the adaptation}
}
\description{
This is the core function that runs the model in our program. Before running this function, we need to specify data, prior, JAGS code, etc. using ipd.data function.
}
