% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpful.functions.R
\name{treatment.effect}
\alias{treatment.effect}
\title{Calculate patient-specific treatment effect}
\usage{
treatment.effect(
  ipd = NULL,
  samples = NULL,
  newpatient = NULL,
  response = "normal",
  quantile = c(0.025, 0.5, 0.975)
)
}
\arguments{
\item{ipd}{ipd object created from ipd.data function}

\item{samples}{mcmc samples found from running `ipd.run`}

\item{newpatient}{Covariate values of patients that you want to predict treatment effect on. Must have length equal to total number of covariates.}

\item{response}{"normal" for continuous outcome and "binomial" for binary outcome. Reports odds ratio for the binary outcome.}

\item{quantile}{quantile for the confidence interval}
}
\description{
Convenient function for calculating the patient-specific treatment effect.
Patient-specific treatment effect includes the main effect of treatment and 
treatment-covariate interaction effect (i.e. effect modification)
}
