% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpful.functions.R
\name{treatment.effect}
\alias{treatment.effect}
\title{Calculate patient-specific treatment effect}
\usage{
treatment.effect(
  result = NULL,
  newpatient = NULL,
  type = "continuous",
  quantile = c(0.025, 0.5, 0.975),
  coef = NULL,
  cov = NULL,
  treatment.covariate.names = NULL
)
}
\arguments{
\item{result}{mcmc samples found from running `ipd.run`}

\item{newpatient}{Covariate values of patients that you want to predict treatment effect on. Must have length equal to total number of covariates. Can also be a matrix of dimension number of different patient groups x number of covariates}

\item{type}{"continuous" for continuous outcome and "binary" for binary outcome. Reports odds ratio for the binary outcome.}

\item{quantile}{quantile for the confidence interval}

\item{coef}{(only for a frequentist method) vector of coefficient values for treatment effect and treatment-covariate interaction effect.}

\item{cov}{(only for a frquentist method) Covariance matrix of the parameters specified in `coef`. Used to find confidence interval.}

\item{treatment.covariate.names}{(only for a frquentist method) Index names of treatment and treatment-covariate interactions (treatment index come first)}
}
\description{
Convenient function for calculating the patient-specific treatment effect.
Patient-specific treatment effect includes the main effect of treatment and 
treatment-covariate interaction effect (i.e. effect modification)
}
