% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpful.functions.R
\name{unstandardize_coefficients}
\alias{unstandardize_coefficients}
\title{Revert standardized coefficients to unstandardized coefficients}
\usage{
unstandardize_coefficients(ipd, samples)
}
\arguments{
\item{ipd}{first-stage model object of the two-stage model (i.e. ipdnma.twostage.model)}

\item{samples}{samples from first-stage model of the two-stage model}
}
\description{
This is a convenient function to revert standardized coefficients to unstandardized coefficients. 
This is primarily for in two-stage models.
}
\examples{
#TODO

}
