% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpful.functions.R
\name{unstandardize_coefficients}
\alias{unstandardize_coefficients}
\title{Revert standardized coefficients to unstandardized coefficients}
\usage{
unstandardize_coefficients(ipd, samples)
}
\arguments{
\item{ipd}{ipdma.model or ipdnma.model object created}

\item{samples}{samples from running ipd.run function}
}
\value{
\item{y}{unstandardized y}
\item{Sigma}{unstandardized Sigma}
}
\description{
This is a convenient function to revert standardized coefficients to unstandardized coefficients. 
This is primarily used when aggregating model estimates in the two-stage models.
}
\examples{
ds <- generate_ipdnma_example()
ds1 <- ds[ds$studyid == 1,] 
ipd1 <- with(ds1, ipdnma.model.twostage.first(y = y, treat = treat, 
X = cbind(z1, z2), response = "normal", shrinkage = "SSVS"))
\donttest{
samples1 <- ipd.run(ipd1)
}
unstandardize_coefficients(ipd1, samples1)
}
